+++ src/main/java/com/wintercogs/beyonddimensions/integration/module/ae2/datagen/AE2ModuleRecipeProvider.java
@@ -1,10 +1,10 @@
-package com.wintercogs.beyonddimensions.integration.module.rs.datagen;
+package com.wintercogs.beyonddimensions.integration.module.ae2.datagen;
 
-import com.refinedmods.refinedstorage.common.content.Blocks;
+import appeng.core.definitions.AEBlocks;
 import com.wintercogs.beyonddimensions.common.init.BDItems;
 import com.wintercogs.beyonddimensions.datagen.util.BDRecipeProvider;
 import com.wintercogs.beyonddimensions.integration.OtherModIds;
-import com.wintercogs.beyonddimensions.integration.module.rs.init.RSModuleBlocks;
+import com.wintercogs.beyonddimensions.integration.module.ae2.init.AE2ModuleItems;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.RecipeCategory;
@@ -18,10 +18,9 @@ import java.util.concurrent.CompletableFuture;
 
 import static net.neoforged.neoforge.common.conditions.NeoForgeConditions.modLoaded;
 
-public class RSModuleRecipeProvider extends BDRecipeProvider
+public class AE2ModuleRecipeProvider extends BDRecipeProvider
 {
-
-    public RSModuleRecipeProvider(HolderLookup.Provider registries, RecipeOutput output)
+    protected AE2ModuleRecipeProvider(HolderLookup.Provider registries, RecipeOutput output)
     {
         super(registries, output);
     }
@@ -29,17 +28,17 @@ public class RSModuleRecipeProvider extends BDRecipeProvider
     @Override
     protected void buildRecipes()
     {
-        RecipeOutput compatOutput = this.output.withConditions(modLoaded(OtherModIds.REFINED_STORAGE));
+        RecipeOutput compatOutput = this.output.withConditions(modLoaded(OtherModIds.AE2));
 
-        ShapedRecipeBuilder.shaped(this.items, RecipeCategory.MISC, RSModuleBlocks.RS_NET_PATHWAY.get())
+        ShapedRecipeBuilder.shaped(this.items, RecipeCategory.MISC, AE2ModuleItems.NET_AE_STORAGE_CELL)
                 .pattern("ABA")
-                .pattern("ACA")
-                .pattern("ADA")
-                .define('A', com.refinedmods.refinedstorage.common.content.Items.INSTANCE.getQuartzEnrichedIron())
-                .define('B', BDItems.SPACE_TIME_STABLE_FRAME.get())
-                .define('C', Blocks.INSTANCE.getMachineCasing())
-                .define('D', Items.REDSTONE)
-                .unlockedBy("unlock_rs_net_pathway", has(BDItems.SPACE_TIME_STABLE_FRAME.get()))
+                .pattern("BDB")
+                .pattern("CCC")
+                .define('A', AEBlocks.QUARTZ_GLASS)
+                .define('B', Items.DIAMOND)
+                .define('C', BDItems.SPACE_TIME_BAR)
+                .define('D', BDItems.SPACE_TIME_STABLE_FRAME)
+                .unlockedBy("unlock_net_ae_storage_cell", has(BDItems.SPACE_TIME_BAR))
                 .save(compatOutput);
     }
 
@@ -53,13 +52,13 @@ public class RSModuleRecipeProvider extends BDRecipeProvider
         @Override
         protected @NotNull RecipeProvider createRecipeProvider(HolderLookup.@NotNull Provider provider, @NotNull RecipeOutput output)
         {
-            return new RSModuleRecipeProvider(provider, output);
+            return new AE2ModuleRecipeProvider(provider, output);
         }
 
         @Override
         public @NotNull String getName()
         {
-            return "BeyondDimensions RSModule Recipe Provider";
+            return "BeyondDimensions AE2Module Recipe Provider";
         }
     }
 }