+++ src/main/java/com/wintercogs/beyonddimensions/integration/module/ae2/item/NetAEStorageCell.java
@@ -1,0 +1,30 @@
+package com.wintercogs.beyonddimensions.integration.module.ae2.item;
+
+import com.wintercogs.beyonddimensions.common.item.NetedItem;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.NotNull;
+
+public class NetAEStorageCell extends NetedItem
+{
+    public NetAEStorageCell(Properties properties)
+    {
+        super(properties.stacksTo(1));
+    }
+
+    @Override
+    public @NotNull InteractionResult use(Level level, Player player, InteractionHand usedHand)
+    {
+        super.use(level, player, usedHand);
+        ItemStack itemstack = player.getItemInHand(usedHand);
+        if (usedHand != InteractionHand.MAIN_HAND || player.isShiftKeyDown())
+        {
+            return InteractionResult.FAIL;
+        }
+
+        return InteractionResult.PASS;
+    }
+}