+++ docs/CHANGELOG.md
@@ -1,2 +1,2 @@
-* restore RS support
+* restore RS support (thanks rocascent)
 * update to mc 26.1.2