+++ docs/CHANGELOG.md
@@ -1,1 +1,2 @@
 * restore AE2 support
+* upgrade to neoforge 26.1.2.41-beta